update users controller added exceptions

+++ src/users/users.controller.ts
@@ -1,8 +1,9 @@
-import { Controller, Get, Post, Put, Body, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Put, Body, Param, Delete, HttpCode, HttpStatus } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { ApiTags } from '@nestjs/swagger';
+import { HttpException } from '@nestjs/common';
 
 @ApiTags('ecommerce')
 @Controller('users')
@@ -22,19 +23,21 @@ export class UsersController {
   @Get(':username')
   async findOne(@Param('username') username: string) {
     const res = await this.usersService.findOne(username)
-    if(!res) return { message: `User ${username} not found` }
+    if(!res) throw new HttpException(`Account with username ${username} Not Found`, HttpStatus.NOT_FOUND);
     return res
   }
 
   @Put(':username')
   async UpdateUser(@Param('username') username: string, @Body() updateUserDto: UpdateUserDto) {
     const updatedUser = await this.usersService.updateUserByUsername(username, updateUserDto)
-    if(!updatedUser) return { message: `User ${username} not found`}
+    if(!updatedUser) throw new HttpException(`Account with username ${username} Not Found`, HttpStatus.NOT_FOUND);
     return updatedUser
   }
 
   @Delete(':username')
-  remove(@Param('username') username: string) {
+  async remove(@Param('username') username: string) {
+    const checkUser = await this.usersService.findOne(username)
+    if (!checkUser) throw new HttpException(`Account with username ${username} Not Found`, HttpStatus.NOT_FOUND);
     return this.usersService.remove(username);
   }
 }
